requirements.txt
aiofiles==24.1.0
click==8.2.1
grpcio-tools==1.73.0
pip==25.1.1
psutil==7.0.0
pydantic==2.11.7
pytest-asyncio==1.0.0
structlog==25.4.0